Fix test

+++ src/vs/platform/agentHost/node/commandAutoApprover.ts
@@ -174,6 +174,10 @@ export class CommandAutoApprover extends Disposable {
 		try {
 			const { default: TreeSitter } = (await import('@vscode/tree-sitter-wasm'));
 
+			if (this._store.isDisposed) {
+				return;
+			}
+
 			// Resolve WASM files from node_modules
 			const moduleRoot = URI.joinPath(FileAccess.asFileUri(''), '..', 'node_modules', '@vscode', 'tree-sitter-wasm', 'wasm');
 			const wasmPath = URI.joinPath(moduleRoot, 'tree-sitter.wasm').fsPath;
@@ -184,13 +188,26 @@ export class CommandAutoApprover extends Disposable {
 				}
 			});
 
+			if (this._store.isDisposed) {
+				return;
+			}
+
 			const parser = new TreeSitter.Parser();
 			this._register(toDisposable(() => parser.delete()));
 
 			// Load bash grammar
 			const bashWasmPath = URI.joinPath(moduleRoot, 'tree-sitter-bash.wasm').fsPath;
 			const bashWasm = await fs.promises.readFile(bashWasmPath);
+
+			if (this._store.isDisposed) {
+				return;
+			}
+
 			const bashLanguage = await TreeSitter.Language.load(new Uint8Array(bashWasm.buffer, bashWasm.byteOffset, bashWasm.byteLength));
+
+			if (this._store.isDisposed) {
+				return;
+			}
 
 			this._parser = parser;
 			this._bashLanguage = bashLanguage;
